(feat) added webpack, moved auth helpers to auth folder
